+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`2155` Users can now additionally filter the uniswap liquidity pools using a pool filter.
 * :feature:`1865` Users will now see an explanation of the current stage of the profit/loss report's progress along with the completion percentage.
 * :bug:`2135` Users will now properly see the correct accounting settings when creating a profit/loss report.